Moved some functionalities to utils.ts
