+++ src/pages/RestaurantsPage/RestaurantsPage.tsx
@@ -5,13 +5,14 @@ import { Fade } from "react-awesome-reveal";
 import "./RestaurantsPage.scss";
 import { RootState, AppDispatch } from "../../redux-toolkit/store";
 import { fetchRestaurantsPageData } from "../../redux-toolkit/thunks/restaurantsPageThunk";
-import Card from "../../components/Card/Card";
+import { filterNewRestaurants } from "../../utils/redux-utils";
 import { CardProps } from "../../models/types";
+
+import Card from "../../components/Card/Card";
 
 const RestaurantsPage = () => {
   const dispatch = useDispatch<AppDispatch>();
   const [activeFilter, setActiveFilter] = useState("All");
-
   const { restaurants } = useSelector((state: RootState) => state.restaurantsPage);
 
   useEffect(() => {
@@ -24,28 +25,15 @@ const RestaurantsPage = () => {
   }, []);
 
   const getFilteredCards = useCallback(() => {
-    const threeDaysAgo = new Date();
-    threeDaysAgo.setDate(threeDaysAgo.getDate() - 5);
-
     switch (activeFilter) {
       case "New":
-        return restaurants.cards.filter((restaurant) => {
-          if (!restaurant._id) return false;
-          const creationDate = getDateFromObjectId(restaurant._id);
-          return creationDate > threeDaysAgo;
-        });
+        return filterNewRestaurants(restaurants.cards, 5);
       case "Most Popular":
         return restaurants.cards.filter((restaurant) => restaurant.isPopular);
       default:
         return restaurants.cards;
     }
   }, [activeFilter, restaurants.cards]);
-
-  const getDateFromObjectId = (objectId: string): Date => {
-    const timestamp = objectId.substring(0, 8);
-    const date = new Date(parseInt(timestamp, 16) * 1000);
-    return date;
-  };
 
   return (
     <>
